Fix: remove metrics command
