+++ references/cli/status.go
@@ -24,6 +24,8 @@ import (
 	"strings"
 	"time"
 
+	"k8s.io/client-go/rest"
+
 	"github.com/fatih/color"
 	"github.com/olekukonko/tablewriter"
 	"github.com/pkg/errors"
@@ -53,6 +55,7 @@ import (
 	cmdutil "github.com/oam-dev/kubevela/pkg/utils/util"
 	types2 "github.com/oam-dev/kubevela/pkg/velaql/providers/query/types"
 	"github.com/oam-dev/kubevela/references/appfile"
+	references "github.com/oam-dev/kubevela/references/common"
 )
 
 // HealthStatus represents health status strings.
@@ -195,7 +198,7 @@ func NewAppStatusCommand(c common.Args, order string, ioStreams cmdutil.IOStream
 	cmd.Flags().BoolVarP(&detail, "detail", "d", false, "display more details in the application like input/output data in context. Note that if you want to show the realtime details of application resources, please use it with --tree")
 	cmd.Flags().StringP("detail-format", "", "inline", "the format for displaying details, must be used with --detail. Can be one of inline, wide, list, table, raw.")
 	cmd.Flags().StringVarP(&outputFormat, "output", "o", "", "raw Application output format. One of: (json, yaml, jsonpath)")
-	cmd.Flags().BoolP("metrics", "m", false, "show resource num nad resource metrics of the application")
+	cmd.Flags().BoolP("metrics", "m", false, "show resource quota and consumption metrics of the application")
 	addNamespaceAndEnvArg(cmd)
 	return cmd
 }
@@ -568,4 +571,34 @@ func printRawApplication(ctx context.Context, c common.Args, format string, out 
 
 	_, err = out.Write([]byte(str))
 	return err
+}
+
+// printMetrics prints the resource num and resource metrics of an application
+func printMetrics(c client.Client, conf *rest.Config, appName, appNamespace string) error {
+	app := new(v1beta1.Application)
+	err := c.Get(context.Background(), client.ObjectKey{
+		Name:      appName,
+		Namespace: appNamespace,
+	}, app)
+	if err != nil {
+		return err
+	}
+	metrics, err := references.LoadApplicationMetrics(c, conf, app)
+	if err != nil {
+		return err
+	}
+	fmt.Println()
+	fmt.Printf("Kubernetes Resources created:\n")
+	fmt.Printf("    * Number of Pods: %d\n", metrics.Resource.PodNum)
+	fmt.Printf("    * Number of Containers: %d\n", metrics.Resource.ContainerNum)
+	fmt.Printf("    * Number of Managed Resource: %d\n", metrics.Resource.SubresourceNum)
+	fmt.Printf("    * Number of Nodes: %d\n", metrics.Resource.NodeNum)
+	fmt.Printf("    * Number of Clusters: %d\n", metrics.Resource.ClusterNum)
+	fmt.Println()
+	fmt.Printf("Underlying Physical Resoures consumed:\n")
+	fmt.Printf("    * Total CPU(cores): %d m\n", metrics.Status.CPU)
+	fmt.Printf("    * Total MEMORY(bytes): %d Mi\n", metrics.Status.Memory)
+	fmt.Printf("    * Total Storage(bytes): %d Gi\n", metrics.Status.Storage)
+	fmt.Println()
+	return nil
 }
